pyproject.toml
[tool.poetry]
name = "langserve-replit"
version = "0.1.0"
description = ""
authors = ["Donatien Thorez"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.14"
python-dotenv = "1.0.0"
openai = "^1.40.0"
sse-starlette = "1.8.2"
pydantic = "1.10.13"
uvicorn = "0.25.0"
fastapi = "0.108.0"
libmagic = "^1.0"
unstructured = {extras = ["pdf"], version = "^0.15.1"}
transformers = "^4.31.0"
datasets = "^2.14.4"
torch = "^2.0.1"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
